Redeposit Fix ,lot's of Speed and Time Remaining!
- The time progress of the simulation and remaining is now printed to
the console

- Added new gravity centres and masses

- fixed collision detection model. (1) Previously, particles were
sometimes added to a neighboring facet instead of the one they actually
hit. (2) Sometimes a collision wasn't detected at all.

- after spawning particles, the model now checks wether particles were
lifted from the surface or if the gas drag was not enough to do so.
Previously, particles not lifted were counted as redeposit. Now they are
just deleted from the particle pool.

- New approach to ensure integration time step is short enough for
particles just spawned:  After spawning, particles are integrated with a
short time step before being integrated together with the rest of the
particles with the big, standard time step. Previously, all particles
below a given altitude were integrated with a smaller time step. This
was extremely inefficient because this covers all particles in that
region and also for the full duration of the big time step. However it
showed that the small time step is only necessary for the first few
seconds after a particles spawned. Therefore the new approach was
implemented.

- Large parts of the code are now arranged more clearly. Many of the
global variables are gone.

- data output is now in .mat format and will be stored into a folder

- The model now also stores the index of the spawn plate of every
particle (SPI)

- The random node creation now only places ice tiles on plates that are
at least twice the area of the tile

- Warnings for high particle production rate are now gone. Instead the
current particle production rate is always displayed in the console

- Added a plot function to visualize redepositioning